+++ basement/source/Basement.csproj
@@ -57,7 +57,7 @@
     <Compile Include="..\..\GpsConsts.cs">
       <Link>Properties\GpsConsts.cs</Link>
     </Compile>
-    <Compile Include="Additions\GoogleApiClientAdditions.cs" />
+    <Compile Include="Additions\Manifest.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="Additions\AboutAdditions.txt" />
